Moving to the GA version of the EventGrid API
